♻️ refactor(recally): extract hooks and components from 932-line monolith
Split RecallyPage.tsx into focused, single-responsibility files:

- hooks/useRecallyFilters.ts — filter state + articles/digest queries + computed tag data

- hooks/useRecallyMutations.ts — toast + update/delete mutations

- hooks/useRecallyFeeds.ts — feed state + create/poll mutations

- components/RecallySidebar.tsx — sidebar with Views/Refine/Feeds

- components/RecallyArticleList.tsx — center panel with article list + stats

- components/RecallyReader.tsx — reader panel with AI summary + auto-mark-read

- components/ArticleCard, FilterChip, StatusBadge, ToastAlert, SidebarNavItem, StatCard

- constants.ts — shared constants + formatSavedAt helper

RecallyPage.tsx is now a ~100-line orchestrator that composes the pieces.
