+++ web/src/features/recally/components/SidebarNavItem.tsx
@@ -1,0 +1,27 @@
+export function SidebarNavItem({
+  label,
+  count,
+  active,
+  onClick,
+}: {
+  label: string;
+  count?: number;
+  active: boolean;
+  onClick: () => void;
+}) {
+  return (
+    <button
+      onClick={onClick}
+      className={`flex w-full items-center justify-between rounded-md px-2 py-1.5 text-sm transition-colors ${
+        active
+          ? "bg-accent font-medium text-foreground"
+          : "text-muted-foreground hover:bg-accent/50 hover:text-foreground"
+      }`}
+    >
+      <span>{label}</span>
+      {count !== undefined && (
+        <span className="font-mono text-xs text-muted-foreground">{count}</span>
+      )}
+    </button>
+  );
+}
