+++ web/src/features/recally/components/StatCard.tsx
@@ -1,0 +1,8 @@
+export function StatCard({ value, label }: { value: number; label: string }) {
+  return (
+    <div className="rounded-md border border-border bg-card px-3 py-2 shadow-sm">
+      <div className="font-mono text-lg font-semibold text-foreground">{value}</div>
+      <div className="text-xs text-muted-foreground">{label}</div>
+    </div>
+  );
+}
